tehty yksinekrtaisempi tekoäly monimutkaisemman kaveriksi
