+++ Tiralabra_maven/src/main/java/com/mycompany/tiralabra_maven/player/SimpleAi.java
@@ -1,0 +1,66 @@
+package com.mycompany.tiralabra_maven.player;
+
+/**
+ *
+ * @author Joel Nummelin
+ */
+public class SimpleAi implements Ai {
+
+    private int LastMove;
+    private int[] opponentsMoves;
+
+    public SimpleAi() {
+        this.LastMove = -2;
+        this.opponentsMoves = new int[3];
+    }
+
+    @Override
+    public int determineMove() {
+        double[] ds = new double[3];
+
+        for (int i = 0; i < 3; i++) {
+            ds[i] = (Math.random() * (opponentsMoves[i] + 3));
+        }
+
+        int x = 0;
+
+        for (int i = 1; i < 3; i++) {
+            if (ds[i] > ds[x]) {
+                x = i;
+            }
+        }
+        LastMove = x;
+        return x;
+    }
+
+    @Override
+    public void update(int result) {
+        if (result == 0) {
+            opponentsMoves[LastMove]++;
+            return;
+        }
+
+        if (LastMove == 0) {
+            if (result == -1) {
+                opponentsMoves[1]++;
+            }
+            if (result == 1) {
+                opponentsMoves[2]++;
+            }
+        } else if (LastMove == 1) {
+            if (result == 1) {
+                opponentsMoves[0]++;
+            }
+            if (result == -1){
+                opponentsMoves[2]++;
+            }
+        } else if (LastMove == 2) {
+            if (result == 1){
+                opponentsMoves[1]++;
+            }
+            if (result == -1){
+                opponentsMoves[0]++;
+            }
+        }
+    }
+}
